PruefProz funktioniert erstmal
git-svn-id: svn://ntoserver/Datenverteiler/VRZ3/trunk/de.bsvrz.dua.langfehlerlve@8816 40c0be50-5d29-0410-901d-fb59edb5cafc

+++ junit/de/bsvrz/dua/langfehlerlve/AbstraktAtgUeberwacher.java
@@ -51,7 +51,7 @@ abstract class AbstraktAtgUeberwacher implements ClientReceiverInterface {
 	 * Der Wert mit dem der erste echte Nutzwert gegengeprueft werden
 	 * soll.
 	 */
-	int wert = DUAKonstanten.NICHT_ERMITTELBAR_BZW_FEHLERHAFT;
+	long wert = DUAKonstanten.NICHT_ERMITTELBAR_BZW_FEHLERHAFT;
 	
 	/**
 	 * Erstes empfangenes Nutzdatum.
